apps/agent/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
# starlette==0.36.3
openai==1.58.1
llama-index==0.12.8
colorama==0.4.6
requests>=2.32.5
python-dotenv==1.0.1
pydantic ==2.10.4
gunicorn==22.0.0
dashscope>=1.22.1
# gevent>=23.9.1
ragflow-sdk==0.23.1
peewee==3.17.3
pymysql>=1.1.1,<2.0.0
openpyxl

# Integration langfuse
langfuse>=3.0.0,<4.0.0

# 接入 duckduckgo-search, 用于网络搜索工具
ddgs==9.10.0

# Integration langgraph
langgraph==1.0.5
langchain-community==0.4.1
langchain==1.2.0
# ipython
python-multipart==0.0.21

# PostgreSQL support for LangGraph Memory
psycopg[binary]>=3.1.0
langgraph-checkpoint-postgres>=1.0.0

# Token counting for accurate pricing
tiktoken>=0.5.0
